+++ lib/src/application.dart
@@ -1,21 +1,23 @@
 import 'package:catechism/src/features/configuration/data/configuration_provider.dart';
 import 'package:catechism/src/features/configuration/data/titles_provider.dart';
 import 'package:catechism/src/features/configuration/domain/configuration.dart';
+import 'package:catechism/src/features/configuration/domain/localized_titles.dart';
 import 'package:catechism/src/features/custom_pages/data/custom_page_provider.dart';
-import 'package:catechism/src/features/custom_pages/domain/custom_page_data.dart';
+import 'package:catechism/src/features/custom_pages/domain/localized_custom_pages.dart';
 import 'package:catechism/src/features/questions/data/question_provider.dart';
-import 'package:catechism/src/features/questions/domain/question.dart';
+import 'package:catechism/src/features/questions/domain/localized_questions.dart';
 import 'package:catechism/src/routing/app_router.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 /// The CatechismApp class is used to create a catechism app.
 class CatechismApp extends StatelessWidget {
   /// The questions property is used to create the list of questions.
-  final List<Question> questions;
+  final LocalizedQuestions questions;
 
   /// The titles property is used to create the list of titles.
-  final List<String> titles;
+  final LocalizedTitles titles;
 
   /// The themeData property is used to create the theme data.
   final ThemeData themeData;
@@ -28,7 +30,7 @@ class CatechismApp extends StatelessWidget {
   final List<Locale> supportedLocales;
 
   /// The customPages property is used to create the list of info screens.
-  final List<CustomPageData> customPages;
+  final LocalizedCustomPages customPages;
 
   /// The CatechismApp constructor is used to create a catechism app.
   CatechismApp({
@@ -39,11 +41,11 @@ class CatechismApp extends StatelessWidget {
     required this.supportedLocales,
     required this.customPages,
   }) {
-    questionProvider = Provider<List<Question>>((ref) => questions);
-    titlesProvider = Provider<List<String>>((ref) => titles);
+    questionProvider = Provider<LocalizedQuestions>((ref) => questions);
+    titlesProvider = Provider<LocalizedTitles>((ref) => titles);
     configurationProvider =
         Provider<CatechismConfiguration>((ref) => configuration);
-    customPageProvider = Provider<List<CustomPageData>>((ref) => customPages);
+    customPageProvider = Provider<LocalizedCustomPages>((ref) => customPages);
   }
 
   /// The build method is used to create the widget.
@@ -52,9 +54,14 @@ class CatechismApp extends StatelessWidget {
     final goRouter = CatechismRouter(customPages: customPages).getRouter();
 
     return MaterialApp.router(
-      title: titles.join(" "),
+      title: titles.defaultTitles.join(" "),
       theme: themeData,
       supportedLocales: supportedLocales,
+      localizationsDelegates: [
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+      ],
       routerConfig: goRouter,
       debugShowCheckedModeBanner: false,
     );