+++ lib/src/features/custom_pages/data/custom_page_provider.dart
@@ -1,4 +1,4 @@
-import 'package:catechism/catechism.dart';
+import 'package:catechism/src/features/custom_pages/domain/localized_custom_pages.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-Provider<List<CustomPageData>>? customPageProvider;
+Provider<LocalizedCustomPages>? customPageProvider;