+++ lib/src/features/questions/presentation/screens/question_details_screen.dart
@@ -1,7 +1,7 @@
+import 'package:catechism/src/common_widgets/top_bar.dart';
 import 'package:catechism/src/features/configuration/data/titles_provider.dart';
 import 'package:catechism/src/features/questions/data/question_provider.dart';
 import 'package:catechism/src/features/questions/presentation/widgets/question_details.dart';
-import 'package:catechism/src/common_widgets/top_bar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -25,19 +25,24 @@ class QuestionDetailsScreen extends ConsumerWidget {
     final PageController controller = PageController(
       initialPage: questionId - 1,
     );
+    final locale = Localizations.localeOf(context);
+    final languageCode = locale.languageCode;
     final titles = ref.watch(titlesProvider!);
+    final localeTitles = titles.getTitles(languageCode);
 
     return Scaffold(
       appBar: TopBar(
-        titles,
+        localeTitles,
         hasBackButton: true,
       ),
       body: Consumer(builder: (context, ref, _) {
         final questions = ref.watch(questionProvider!);
+        final localeQuestions = questions.getQuestions(languageCode);
+
         return PageView(
           scrollDirection: Axis.horizontal,
           controller: controller,
-          children: questions
+          children: localeQuestions
               .map((e) => QuestionDetails(
                     question: e,
                   ))